Changes to Flod 4.1 as to be exported by ax3
Changed made to the Flod 4.1 AS3 package to name it exportable with ax3. Mostly removing multiple var statements in classes and adding extra brackets to if/else statements that ax3 was struggling with and a couple spots where semicolons where missing from line ends.
Also a spot fix to some 0 padded integers that where confusing ax3 in D1Player.