A bunch of fixed made to the ax3 output to get it to compile
A bunch of fixed made to the ax3 output to get it to compile. Because I'm not including ASCompat, flash.Vector project wide or ReverseIntIterator there where a bunch of spot fixes as well some some small fixes where variables where defined as dyn when a specific haxe type would work.